refactor(l1): split staffing lifecycle operations
