+++ simulation/l1_staffing.py
@@ -7,19 +7,22 @@ import copy
 from dataclasses import asdict, dataclass, field
 from typing import Any, Dict, List, Optional, Tuple
 
+from l1_staffing_lifecycle import (
+    mark_persona_resolved,
+    record_personnel_observation,
+    retire_staffing_seat,
+    transfer_personnel_from_orion,
+)
 from l1_staffing_validation import (
     ALLOWED_OBSERVATION_FIELDS,
     PROHIBITED_PERSONA_FIELDS,
     action_personnel_ids as _action_personnel_ids,
     complement_delta_from_actions as _complement_delta_from_actions,
-    lifecycle_action as _lifecycle_action,
     require_nonnegative_integers as _require_nonnegative_integers,
     require_nonnegative_numeric as _require_nonnegative_numeric,
     resolved_personnel_ids as _resolved_personnel_ids,
     stable_id as _stable_id,
     validate_action_subject as _validate_action_subject,
-    validate_lifecycle_inputs as _validate_lifecycle_inputs,
-    validate_observation_payload as _validate_observation_payload,
 )
 
 
@@ -437,145 +440,6 @@ def apply_staffing_decision(
     staffing.actions.append(action)
     if decision.action_type in {"external_hire", "transfer_to_orion"}:
         staffing.human_complement_delta += 1
-    staffing.validate()
-    return copy.deepcopy(action)
-
-
-def transfer_personnel_from_orion(
-    staffing: StaffingRunState,
-    personnel_id: str,
-    *,
-    provenance: str,
-    rationale: str,
-    receipt_id: str,
-    tick: int,
-) -> Dict[str, Any]:
-    """Record an outbound transfer without erasing the personnel identity."""
-    _validate_lifecycle_inputs(provenance, rationale, receipt_id)
-    personnel = staffing.personnel.get(personnel_id)
-    if personnel is None:
-        raise ValueError("outbound transfer references unknown personnel")
-    if personnel.employment_status in {"departed", "off_station"}:
-        raise ValueError("personnel is already off Orion")
-    if personnel.employment_status in {"contractor", "visitor"}:
-        raise ValueError(
-            "non-complement personnel departure cannot change human crew complement"
-        )
-    before = asdict(personnel)
-    personnel.employment_status = "departed"
-    personnel.shift_status = "transferred_off_station"
-    personnel.workload_status = "not_assigned"
-    action = _lifecycle_action(
-        action_type="transfer_from_orion",
-        subject_id=personnel_id,
-        receipt_id=receipt_id,
-        provenance=provenance,
-        rationale=rationale,
-        tick=tick,
-        before=before,
-        after=asdict(personnel),
-    )
-    staffing.actions.append(action)
-    staffing.human_complement_delta -= 1
-    staffing.validate()
-    return copy.deepcopy(action)
-
-
-def retire_staffing_seat(
-    staffing: StaffingRunState,
-    staffing_seat: str,
-    *,
-    provenance: str,
-    rationale: str,
-    receipt_id: str,
-    tick: int,
-) -> Dict[str, Any]:
-    """Retire a vacant organizational seat while preserving its provenance."""
-    _validate_lifecycle_inputs(provenance, rationale, receipt_id)
-    seat = staffing.seats.get(staffing_seat)
-    if seat is None:
-        raise ValueError("seat retirement references unknown staffing seat")
-    if seat.status == "retired":
-        raise ValueError("staffing seat is already retired")
-    if any(
-        record.staffing_seat == staffing_seat
-        and record.employment_status not in {"departed", "off_station"}
-        for record in staffing.personnel.values()
-    ):
-        raise ValueError("occupied staffing seat cannot be retired")
-    before = asdict(seat)
-    seat.status = "retired"
-    seat.retirement_provenance = provenance
-    action = _lifecycle_action(
-        action_type="seat_retirement",
-        subject_id=staffing_seat,
-        receipt_id=receipt_id,
-        provenance=provenance,
-        rationale=rationale,
-        tick=tick,
-        before=before,
-        after=asdict(seat),
-    )
-    staffing.actions.append(action)
-    staffing.validate()
-    return copy.deepcopy(action)
-
-
-def record_personnel_observation(
-    staffing: StaffingRunState,
-    personnel_id: str,
-    observations: Dict[str, str],
-    *,
-    provenance: str,
-) -> PersonnelRecord:
-    """Add observed traits without synthesizing biography or canon persona."""
-    record = staffing.personnel.get(personnel_id)
-    if record is None:
-        raise ValueError("staffing observation references unknown personnel")
-    _validate_observation_payload(observations, provenance)
-    record.observed_traits.update(copy.deepcopy(observations))
-    if provenance not in record.observation_provenance:
-        record.observation_provenance.append(provenance)
-    if record.persona_resolution == "minimal":
-        record.persona_resolution = "partially_observed"
-    record.validate()
-    staffing.validate()
-    return copy.deepcopy(record)
-
-
-def mark_persona_resolved(
-    staffing: StaffingRunState,
-    personnel_id: str,
-    *,
-    receipt_id: str,
-    tick: int,
-) -> Dict[str, Any]:
-    """Mark a run-state persona resolved only after observation and review."""
-    record = staffing.personnel.get(personnel_id)
-    if record is None:
-        raise ValueError("persona resolution references unknown personnel")
-    if not record.observed_traits:
-        raise ValueError("persona resolution requires prior observations")
-    if record.persona_resolution == "persona_resolved_run_state":
-        raise ValueError("personnel persona is already resolved in this run")
-    record.persona_resolution = "persona_resolved_run_state"
-    staffing.persona_resolved_delta += 1
-    action = {
-        "action_id": _stable_id(
-            "staffing-action",
-            personnel_id,
-            "persona-resolution",
-            str(tick),
-        ),
-        "action_type": "persona_resolution",
-        "demand_id": "progressive_persona_resolution",
-        "personnel_id": personnel_id,
-        "receipt_id": receipt_id,
-        "tick": tick,
-        "rationale": "Observed run-state traits reviewed for progressive resolution.",
-        "canon_status": "run_state_not_canon_promotion",
-    }
-    staffing.actions.append(action)
     staffing.validate()
     return copy.deepcopy(action)
